src/aiops-lambda/requirements.txt
requests==2.31.0
boto3==1.34.0
google-generativeai==0.3.2
